refactor: shared/model/type.ts 폴더 이동
