+++ src/shared/model/type.ts
