+++ src/shared/ui/Button/model/type.ts
@@ -1,5 +1,5 @@
 import { cva, VariantProps } from "class-variance-authority"
-import { ClassNameProps } from "../../type"
+import { ClassNameProps } from "../../../model/type"
 
 export const buttonVariants = cva(
   "inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background",
