+++ src/shared/ui/Card/model/type.ts
@@ -1,3 +1,3 @@
-import { ClassNameProps } from "../../type"
+import { ClassNameProps } from "../../../model/type"
 
 export interface CardProps extends React.HTMLAttributes<HTMLDivElement>, ClassNameProps {}
