+++ src/shared/ui/Input/model/type.ts
@@ -1,3 +1,3 @@
-import { ClassNameProps } from "../../type"
+import { ClassNameProps } from "../../../model/type"
 
 export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement>, ClassNameProps {}
