+++ src/shared/ui/Textarea/model/type.ts
@@ -1,3 +1,3 @@
-import { ClassNameProps } from "../../type"
+import { ClassNameProps } from "../../../model/type"
 
 export interface TextareaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement>, ClassNameProps {}
